move options to v-upload
